# client/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Financial Document Analyzer API",
      "description": "FastAPI application with JWT auth, document management, and CrewAI agents",
      "version": "2.0.0"
    },
    "paths": {
...
      "/analysis/comprehensive": {
        "post": {
          "tags": [
            "analysis"
          ],
          "summary": "Analyze Comprehensive",
          "description": "Comprehensive financial document analysis - runs in background",
          "operationId": "analyze_comprehensive_analysis_comprehensive_post",
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/Body_analyze_comprehensive_analysis_comprehensive_post"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ]
        }
      },
      "/analysis/investment": {
        "post": {
          "tags": [
            "analysis"
          ],
          "summary": "Analyze Investment",
          "description": "Investment-focused financial document analysis",
          "operationId": "analyze_investment_analysis_investment_post",
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/Body_analyze_investment_analysis_investment_post"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ]
        }
      },
      "/analysis/risk": {
        "post": {
          "tags": [
            "analysis"
          ],
          "summary": "Analyze Risk",
          "description": "Risk-focused financial document analysis",
          "operationId": "analyze_risk_analysis_risk_post",
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/Body_analyze_risk_analysis_risk_post"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ]
        }
      },
      "/analysis/verify": {
        "post": {
          "tags": [
            "analysis"
          ],
          "summary": "Verify Document",
          "description": "Verify if the document is a valid financial record",
          "operationId": "verify_document_analysis_verify_post",
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/Body_verify_document_analysis_verify_post"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ]
        }
      },
      "/analysis/types": {
        "get": {
          "tags": [
            "analysis"
          ],
          "summary": "Get Analysis Types",
          "description": "Get available analysis types",
          "operationId": "get_analysis_types_analysis_types_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ]
        }
      },
...
      "/reports/{report_id}/download": {
        "get": {
          "tags": [
            "reports"
          ],
          "summary": "Download Report",
          "description": "Download an analysis report file",
          "operationId": "download_report_reports__report_id__download_get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "report_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "description": "Report ID",
                "title": "Report Id"
              },
              "description": "Report ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Body_analyze_comprehensive_analysis_comprehensive_post": {
          "properties": {
            "query": {
              "type": "string",
              "title": "Query",
              "default": "Analyze this financial document for comprehensive insights"
            },
            "file": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "binary"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File"
            },
            "document_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          },
          "type": "object",
          "title": "Body_analyze_comprehensive_analysis_comprehensive_post"
        },
        "Body_analyze_investment_analysis_investment_post": {
          "properties": {
            "query": {
              "type": "string",
              "title": "Query",
              "default": "Analyze this financial document for investment opportunities"
            },
            "file": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "binary"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File"
            },
            "document_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          },
          "type": "object",
          "title": "Body_analyze_investment_analysis_investment_post"
        },
        "Body_analyze_risk_analysis_risk_post": {
          "properties": {
            "query": {
              "type": "string",
              "title": "Query",
              "default": "Analyze this financial document for risk assessment"
            },
            "file": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "binary"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File"
            },
            "document_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          },
          "type": "object",
          "title": "Body_analyze_risk_analysis_risk_post"
        },
...
        "Body_verify_document_analysis_verify_post": {
          "properties": {
            "query": {
              "type": "string",
              "title": "Query",
              "default": "Verify if this is a valid financial document"
            },
            "file": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "binary"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File"
            },
            "document_id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          },
          "type": "object",
          "title": "Body_verify_document_analysis_verify_post"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "HTTPBearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    }
  }